+++ lib/src/base.dart
@@ -2,23 +2,8 @@ part of 'database.dart';
 
 typedef InAppValue = Object?;
 typedef InAppDocument = Map<String, InAppValue>;
-
-class InAppDocumentId {
-  final String id;
-
-  const InAppDocumentId(this.id);
-
-  @override
-  int get hashCode => id.hashCode;
-
-  @override
-  String toString() => id;
-
-  @override
-  bool operator ==(Object other) {
-    return super == other && hashCode == other.hashCode;
-  }
-}
+typedef InAppDatabaseReader = Future<String?> Function(String key);
+typedef InAppDatabaseWriter = Future<bool> Function(String key, String? value);
 
 class InAppSetOptions {
   final bool merge;
@@ -26,4 +11,22 @@ class InAppSetOptions {
   const InAppSetOptions({
     this.merge = false,
   });
+}
+
+enum InAppReadType {
+  collection,
+  document;
+
+  bool get isCollection => this == collection;
+
+  bool get isDocument => this == document;
+}
+
+enum InAppWriteType {
+  collection,
+  document;
+
+  bool get isCollection => this == collection;
+
+  bool get isDocument => this == document;
 }