+++ lib/src/counter.dart
@@ -1,35 +1,31 @@
 part of 'database.dart';
 
 class InAppCounterReference extends InAppReference {
-  final InAppCollectionReference parent;
-  final String collectionPath;
-  final String collectionId;
+  final InAppCollectionReference _p;
 
   const InAppCounterReference({
     required super.db,
     required super.reference,
-    required this.collectionPath,
-    required this.collectionId,
-    required this.parent,
-  });
+    required InAppCollectionReference parent,
+  }) : _p = parent;
 
   Future<InAppCounterSnapshot> get() {
-    return parent.get().then((value) {
-      return InAppCounterSnapshot(collectionId, value.docs.length);
+    return _p.get().then((value) {
+      return InAppCounterSnapshot(_p.id, value.docs.length);
     });
   }
 
   Stream<InAppCounterSnapshot> snapshots() {
     final c = StreamController<InAppCounterSnapshot>();
-    final n = db.addListener(collectionPath);
+    final n = _db._addNotifier(_p.path);
     n.addListener(() {
       c.add(InAppCounterSnapshot(
-        collectionId,
+        _p.id,
         n.value?.docs.length ?? 0,
         n.value?.docChanges.length ?? 0,
       ));
     });
-    Future.delayed(const Duration(seconds: 1)).whenComplete(parent.notify);
+    Future.delayed(const Duration(seconds: 1)).whenComplete(_p._notify);
     return c.stream;
   }
 }