+++ lib/src/database.dart
@@ -4,78 +4,90 @@ import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:in_app_query/in_app_query.dart';
 
+import '../core/field_value.dart';
 import '../core/paging_options.dart';
 
 part 'base.dart';
 part 'collection.dart';
 part 'counter.dart';
 part 'document.dart';
+part 'merger.dart';
 part 'notifier.dart';
-part 'params.dart';
 part 'query.dart';
 part 'reference.dart';
 part 'snapshots.dart';
 
 class InAppDatabase {
   final String name;
-  final InAppDatabaseReader reader;
-  final InAppDatabaseWriter writer;
-  final Map<String, InAppNotifier> notifiers = {};
+  final InAppDatabaseReader _reader;
+  final InAppDatabaseWriter _writer;
+  final Map<String, _InAppNotifier> _notifiers = {};
 
   InAppDatabase._({
     required this.name,
-    required this.reader,
-    required this.writer,
-  });
+    required InAppDatabaseReader reader,
+    required InAppDatabaseWriter writer,
+  })  : _reader = reader,
+        _writer = writer;
 
   static InAppDatabase? _i;
 
   static InAppDatabase get i => _i!;
 
+  static InAppDatabase get I => i;
+
+  static InAppDatabase get instance => i;
+
   static Future<InAppDatabase> init({
+    String? name,
     required InAppDatabaseReader reader,
     required InAppDatabaseWriter writer,
   }) async {
     _i ??= InAppDatabase._(
-      name: "__in_app_database__",
+      name: name ?? "__in_app_database__",
       reader: reader,
       writer: writer,
     );
     return _i!;
   }
 
-  InAppQueryNotifier addListener(
-    String reference, [
-    InAppQuerySnapshot? value,
-  ]) {
-    final i = notifiers[reference];
-    if (i == null) notifiers[reference] = InAppQueryNotifier(value);
-    final x = i ?? notifiers[reference];
-    return x is InAppQueryNotifier ? x : InAppQueryNotifier(value);
-  }
-
-  InAppDocumentNotifier addChildListener(
-    String reference,
-    String id, [
-    InAppDocumentSnapshot? value,
-  ]) {
-    return addListener(reference).set(id, value);
-  }
-
   InAppQueryReference ref(String field) {
     return InAppQueryReference(
       db: this,
       reference: "$name$field",
-      collectionPath: field,
-      collectionId: field,
+      path: field,
+      id: field,
     );
   }
 
   InAppQueryReference collection(String field) => ref(field);
 
-  Future<InAppSnapshot> _r(InAppReadParams r) {
-    return reader(r.collectionPath).then((raw) {
-      final type = r.type;
+  _InAppQueryNotifier _addNotifier(
+    String reference, [
+    InAppQuerySnapshot? value,
+  ]) {
+    final i = _notifiers[reference];
+    if (i == null) _notifiers[reference] = _InAppQueryNotifier(value);
+    final x = i ?? _notifiers[reference];
+    return x is _InAppQueryNotifier ? x : _InAppQueryNotifier(value);
+  }
+
+  _InAppDocumentNotifier _addChildNotifier(
+    String reference,
+    String id, [
+    InAppDocumentSnapshot? value,
+  ]) {
+    return _addNotifier(reference).set(id, value);
+  }
+
+  Future<InAppSnapshot> _r({
+    required InAppReadType type,
+    required String reference,
+    required String collectionPath,
+    required String collectionId,
+    required String documentId,
+  }) {
+    return _reader(collectionPath).then((raw) {
       final value = raw is String ? jsonDecode(raw) : raw;
       if (value is Map) {
         if (type.isCollection) {
@@ -88,9 +100,9 @@ class InAppDatabase {
               })
               .where((i) => i.data != null && i.data!.isNotEmpty)
               .toList();
-          return InAppQuerySnapshot(r.collectionId, data);
+          return InAppQuerySnapshot(collectionId, data);
         } else if (type.isDocument) {
-          final docId = r.documentId;
+          final docId = documentId;
           final raw = value[docId];
           final doc = raw is String
               ? jsonDecode(raw)
@@ -107,55 +119,26 @@ class InAppDatabase {
     });
   }
 
-  Future<bool> _w(InAppWriteParams w) async {
-    return reader(w.collectionPath).then((value) {
+  Future<bool> _w({
+    required InAppWriteType type,
+    required String reference,
+    required String collectionPath,
+    required String collectionId,
+    required String documentId,
+    InAppDocument? value,
+  }) {
+    return _reader(collectionPath).then((value) {
       final raw = value is String ? jsonDecode(value) : null;
       final base = raw is Map ? raw : {};
-      final data = w.value;
-      final id = w.documentId;
+      final data = value == null ? null : jsonEncode(value);
+      final id = documentId;
       if (data != null) {
         base[id] = data;
       } else {
         base.remove(id);
       }
       final body = jsonEncode(base);
-      return writer(w.collectionPath, body);
+      return _writer(collectionPath, body);
     });
-  }
-
-  Future<InAppSnapshot> read({
-    required InAppReadType type,
-    required String reference,
-    required String collectionPath,
-    required String collectionId,
-    required String documentId,
-  }) {
-    return _r(InAppReadParams(
-      reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
-      documentPath: "$collectionPath/$documentId",
-      documentId: documentId,
-      type: type,
-    ));
-  }
-
-  Future<bool> write({
-    required InAppWriteType type,
-    required String reference,
-    required String collectionPath,
-    required String collectionId,
-    required String documentId,
-    Object? value,
-  }) {
-    return _w(InAppWriteParams(
-      reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
-      documentPath: "$collectionPath/$documentId",
-      documentId: documentId,
-      value: value == null ? null : jsonEncode(value),
-      type: type,
-    ));
   }
 }