+++ lib/src/query.dart
@@ -1,49 +1,51 @@
 part of 'database.dart';
 
 class InAppQueryReference extends InAppCollectionReference {
-  final List<Query> queries;
-  final List<Selection> selections;
-  final List<Sorting> sorts;
-  final InAppPagingOptions options;
-  final bool counterMode;
+  final List<Query> _q;
+  final List<Selection> _s;
+  final List<Sorting> _o;
+  final InAppPagingOptions _op;
+  final bool _cm;
 
   const InAppQueryReference({
     required super.db,
     required super.reference,
-    required super.collectionPath,
-    required super.collectionId,
-    this.queries = const [],
-    this.selections = const [],
-    this.sorts = const [],
-    this.options = const InAppPagingOptions(),
-    this.counterMode = false,
-  });
+    required super.path,
+    required super.id,
+    List<Query> queries = const [],
+    List<Selection> selections = const [],
+    List<Sorting> sorts = const [],
+    InAppPagingOptions options = const InAppPagingOptions(),
+    bool counterMode = false,
+  })  : _q = queries,
+        _s = selections,
+        _o = sorts,
+        _op = options,
+        _cm = counterMode;
 
   InAppCounterReference count() {
     return InAppCounterReference(
-      db: db,
+      db: _db,
       reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
       parent: this,
     );
   }
 
   InAppQueryReference _limit(int limit, [bool fetchFromLast = false]) {
     return InAppQueryReference(
-      db: db,
+      db: _db,
       reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
-      queries: queries,
-      selections: selections,
-      sorts: sorts,
-      options: options.copy(
+      path: path,
+      id: id,
+      queries: _q,
+      selections: _s,
+      sorts: _o,
+      options: _op.copy(
         initialSize: limit,
         fetchingSize: limit,
         fetchFromLast: fetchFromLast,
       ),
-      counterMode: counterMode,
+      counterMode: _cm,
     );
   }
 
@@ -55,38 +57,38 @@ class InAppQueryReference extends InAppCollectionReference {
     Object field, {
     bool descending = false,
   }) {
-    List<Sorting> sorts = List.from(this.sorts);
+    List<Sorting> sorts = List.from(_o);
     if (field is String) {
       sorts.add(Sorting(field, descending: descending));
     }
     return InAppQueryReference(
-      db: db,
+      db: _db,
       reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
-      queries: queries,
-      selections: selections,
+      path: path,
+      id: id,
+      queries: _q,
+      selections: _s,
       sorts: sorts,
-      options: options,
-      counterMode: counterMode,
+      options: _op,
+      counterMode: _cm,
     );
   }
 
   InAppQueryReference _selection(Object? snapshot, Selections type) {
-    List<Selection> selections = List.from(this.selections);
+    List<Selection> selections = List.from(_s);
     if (snapshot is InAppDocument || snapshot is Iterable<InAppValue>) {
       selections.add(Selection.from(snapshot, type));
     }
     return InAppQueryReference(
-      db: db,
+      db: _db,
       reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
-      queries: queries,
+      path: path,
+      id: id,
+      queries: _q,
       selections: selections,
-      sorts: sorts,
-      options: options,
-      counterMode: counterMode,
+      sorts: _o,
+      options: _op,
+      counterMode: _cm,
     );
   }
 
@@ -104,7 +106,7 @@ class InAppQueryReference extends InAppCollectionReference {
     Iterable<Object?>? whereNotIn,
     bool? isNull,
   }) {
-    List<Query> queries = List.from(this.queries);
+    List<Query> queries = List.from(_q);
     queries.add(Query(
       field,
       isEqualTo: isEqualTo,
@@ -120,15 +122,15 @@ class InAppQueryReference extends InAppCollectionReference {
       isNull: isNull,
     ));
     return InAppQueryReference(
-      db: db,
+      db: _db,
       reference: reference,
-      collectionPath: collectionPath,
-      collectionId: collectionId,
+      path: path,
+      id: id,
       queries: queries,
-      selections: selections,
-      sorts: sorts,
-      options: options,
-      counterMode: counterMode,
+      selections: _s,
+      sorts: _o,
+      options: _op,
+      counterMode: _cm,
     );
   }
 
@@ -166,95 +168,106 @@ class InAppQueryReference extends InAppCollectionReference {
 
   @override
   Future<InAppQuerySnapshot> get() {
-    return super.get().then((raw) {
-      final data = raw.docs.map((e) => e.data ?? {}).toList();
-      QueryBuilder builder = QueryBuilder(data);
-      if (queries.isNotEmpty) {
-        for (var i in queries) {
-          builder = builder.where(
-            i.field,
-            isEqualTo: i.isEqualTo,
-            isNotEqualTo: i.isNotEqualTo,
-            isNull: i.isNull,
-            isGreaterThan: i.isGreaterThan,
-            isGreaterThanOrEqualTo: i.isGreaterThanOrEqualTo,
-            isLessThan: i.isLessThan,
-            isLessThanOrEqualTo: i.isLessThanOrEqualTo,
-            whereIn: i.whereIn,
-            whereNotIn: i.whereNotIn,
-            arrayContains: i.arrayContains,
-            arrayNotContains: i.arrayNotContains,
-            arrayContainsAny: i.arrayContainsAny,
-            arrayNotContainsAny: i.arrayNotContainsAny,
-          );
+    final fetchingSize = _op.fetchingSize ?? 0;
+    final isQuery = _q.isNotEmpty;
+    final isSorting = _o.isNotEmpty;
+    final isSelections = _s.isNotEmpty;
+    final isLimit = fetchingSize > 0;
+    final sortingSize = _o.length;
+    if (isQuery || isSorting || isSelections || isLimit) {
+      return super.get().then((raw) {
+        final data = raw.docs.map((e) => e.data ?? {}).toList();
+        QueryBuilder builder = QueryBuilder(data);
+        if (isQuery) {
+          for (var i in _q) {
+            builder = builder.where(
+              i.field,
+              isEqualTo: i.isEqualTo,
+              isNotEqualTo: i.isNotEqualTo,
+              isNull: i.isNull,
+              isGreaterThan: i.isGreaterThan,
+              isGreaterThanOrEqualTo: i.isGreaterThanOrEqualTo,
+              isLessThan: i.isLessThan,
+              isLessThanOrEqualTo: i.isLessThanOrEqualTo,
+              whereIn: i.whereIn,
+              whereNotIn: i.whereNotIn,
+              arrayContains: i.arrayContains,
+              arrayNotContains: i.arrayNotContains,
+              arrayContainsAny: i.arrayContainsAny,
+              arrayNotContainsAny: i.arrayNotContainsAny,
+            );
+          }
         }
-      }
-      if (sorts.isNotEmpty) {
-        for (var i in sorts) {
-          builder = builder.orderBy(i.field, descending: i.descending);
+        if (isSorting) {
+          for (var i in _o) {
+            builder = builder.orderBy(i.field, descending: i.descending);
+          }
         }
-      }
-      if (selections.isNotEmpty) {
-        for (var i in selections) {
-          final v = i.value;
-          final vs = i.values;
-          if (v is Map<String, dynamic> || vs is Iterable<Object?>) {
-            final value = v is Map<String, dynamic> ? v : <String, dynamic>{};
-            final values = List.from(vs ?? []);
-            switch (i.type) {
-              case Selections.endAt:
-                builder = builder.endAt(values);
-                break;
-              case Selections.endAtDocument:
-                builder = builder.endAtDocument(value);
-                break;
-              case Selections.endBefore:
-                builder = builder.endBefore(values);
-                break;
-              case Selections.endBeforeDocument:
-                builder = builder.endBeforeDocument(value);
-                break;
-              case Selections.startAfter:
-                builder = builder.startAfter(values);
-                break;
-              case Selections.startAfterDocument:
-                builder = builder.startAfterDocument(value);
-                break;
-              case Selections.startAt:
-                builder = builder.startAt(values);
-                break;
-              case Selections.startAtDocument:
-                builder = builder.startAtDocument(value);
-                break;
-              case Selections.none:
-                break;
+        if (isSelections) {
+          for (var i in _s) {
+            final v = i.value;
+            final vs = i.values;
+            if (v is Map<String, dynamic> || vs is Iterable<Object?>) {
+              final value = v is Map<String, dynamic> ? v : <String, dynamic>{};
+              final values = List.from(vs ?? []);
+              switch (i.type) {
+                case Selections.endAt:
+                  builder = builder.endAt(values);
+                  break;
+                case Selections.endAtDocument:
+                  builder = builder.endAtDocument(value);
+                  break;
+                case Selections.endBefore:
+                  builder = builder.endBefore(values);
+                  break;
+                case Selections.endBeforeDocument:
+                  builder = builder.endBeforeDocument(value);
+                  break;
+                case Selections.startAfter:
+                  builder = builder.startAfter(values);
+                  break;
+                case Selections.startAfterDocument:
+                  builder = builder.startAfterDocument(value);
+                  break;
+                case Selections.startAt:
+                  builder = builder.startAt(values);
+                  break;
+                case Selections.startAtDocument:
+                  builder = builder.startAtDocument(value);
+                  break;
+                case Selections.none:
+                  break;
+              }
             }
           }
         }
-      }
-      final fetchingSize = options.fetchingSize;
-      if (fetchingSize != null && fetchingSize > 0) {
-        if (options.fetchFromLast) {
-          builder = builder.limitToLast(fetchingSize);
-        } else {
-          builder = builder.limit(fetchingSize);
+        if (isLimit) {
+          if (_op.fetchFromLast) {
+            builder = builder.limitToLast(fetchingSize);
+          } else {
+            builder = builder.limit(fetchingSize);
+          }
         }
-      }
-      return builder.execute().then((processed) {
-        final docs = processed.map((e) {
-          return InAppDocumentSnapshot(id, e);
-        }).toList();
-        return InAppQuerySnapshot(raw.id, docs);
+        return builder
+            .execute(isSorting ? sortingSize * data.length : 0)
+            .then((processed) {
+          final docs = processed.map((e) {
+            return InAppDocumentSnapshot(_id, e);
+          }).toList();
+          return InAppQuerySnapshot(raw.id, docs);
+        });
       });
-    });
+    } else {
+      return super.get();
+    }
   }
 
   @override
   Stream<InAppQuerySnapshot> snapshots() {
     final c = StreamController<InAppQuerySnapshot>();
-    final n = db.addListener(collectionPath);
+    final n = _db._addNotifier(path);
     n.addListener(() => get().then(c.add));
-    Future.delayed(const Duration(seconds: 1)).whenComplete(notify);
+    Future.delayed(const Duration(seconds: 1)).whenComplete(_notify);
     return c.stream;
   }
 }