+++ lib/src/reference.dart
@@ -2,14 +2,14 @@ part of 'database.dart';
 
 abstract class InAppReference {
   final String reference;
-  final InAppDatabase db;
+  final InAppDatabase _db;
 
   const InAppReference({
     required this.reference,
-    required this.db,
-  });
+    required InAppDatabase db,
+  }) : _db = db;
 
-  String get id => DateTime.now().millisecondsSinceEpoch.toString();
+  String get _id => DateTime.now().millisecondsSinceEpoch.toString();
 
-  String get idField => "_id";
+  String get _idField => "_id";
 }