+++ CHANGELOG.md
@@ -1,3 +1,19 @@
-## 1.1.0
+## 1.2.0
 
-* initial release.
+* stable version
+* Add query mode
+* Add filter mode
+* Add sorting mode
+* Add selection mode
+* Add collection mode
+* Add document mode
+* Add retrieve document mode
+* Add retrieve collection mode
+* Add collection stream mode
+* Add document stream mode
+* Add collection document add mode
+* Add document set mode
+* Add document update mode
+* Add existing document field value update mode
+  Like: arrayUnion, arrayRemove, delete, serverTimestamp, increment
+* Add document delete mode